requirements.txt
aiohttp==3.7.4.post0
alembic==1.5.6
aniso8601==9.0.1
appdirs==1.4.4
astroid==2.5.1
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.6
backoff==1.10.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
blinker==1.4
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
config==0.5.0.post0
crontab==0.22.9
decorator==5.0.7
distlib==0.3.1
distro==1.5.0
et-xmlfile==1.0.1
faker==8.0.0
filelock==3.0.12
flasgger==0.9.5
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Caching==1.10.1
Flask-Cors==3.0.10
Flask-FlatPages==0.7.2
Flask-FlatPages-Pandoc==0.2
Flask-JSON==0.3.4
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Migrate==2.7.0
Flask-MySQL==1.5.2
Flask-RESTful==0.3.8
flask-restful-swagger==0.20.2
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-Simple-GeoIP==0.2.4
Flask-SQLAlchemy==2.4.4
flask-swagger-ui==3.36.0
Flask-Testing==0.8.1
Frozen-Flask==0.15
future==0.18.2
gcloud==0.18.3
geocoder==1.38.1
geoip2==4.1.0
google-api-core==1.26.3
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-storage==1.37.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
h11==0.12.0
httplib2==0.19.0
idna==2.10
importlib-metadata==3.7.3
ipaddress==1.0.23
isort==5.7.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.3
jsonschema==3.2.0
jsonschema2md==0.1.1
lazy-object-proxy==1.5.2
logger==1.4
lxml==4.6.2
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.11.1
maxminddb==2.0.3
mccabe==0.6.1
mistune==0.8.4
multidict==5.1.0
numpy==1.20.2
oauth2client==4.1.3
openpyxl==3.0.7
packaging==20.9
pandas==1.2.3
Password-Validation==0.1.1
pbr==5.5.1
pdoc3==0.9.2
Pillow==8.2.0
pipenv==2020.11.15
priority==1.3.0
protobuf==3.15.6
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycountry==20.7.3
pycparser==2.20
pygeoip==0.3.2
PyJWT==2.0.1
pylint==2.7.2
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.16.0
python-editor==1.0.4
python-geoip==1.2
python-geoip-geolite2==2015.303
python-geoip-python3==1.3
pytz==2021.1
PyYAML==5.4.1
ratelim==0.1.6
requests==2.25.1
rsa==4.7.2
schedule==1.0.0
simple-geoip==0.1.1
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.3.23
stevedore==3.3.0
tabulate==0.8.9
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
urlopen==1.0.0
uvicorn==0.13.4
virtualenv==20.4.2
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
virtualenvwrapper.basic==0.2.2
Werkzeug==0.16.1
wincertstore==0.2
wrapt==1.12.1
xlrd==2.0.1
XlsxWriter==1.4.3
xlwt==1.3.0
yarl==1.6.3
zipp==3.4.1
